python-blog/requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.2.13
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-tinymce4-lite==1.7.4
gunicorn==19.9.0
idna==2.8
isort==4.3.4
jsmin==2.2.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==7.1.0
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.1.1
urllib3==1.24.2
wrapt==1.11.0
python-decouple==3.3